Refactored HTTP request in initial data fetch
Chaining requests did not work with the Axios functions. I'm
not sure if this is a limitation of Promises altogether or if
it's a bug with Axios. Regardless, I would like to strip out Axios
and rewrite the HTTP requests as simple Promise requests without the
use of a third party package. During that refactor, I can test the
chaining and see for certain.